requirements.txt
pyspark
python-restcountries
kafka-python
datetime
requests
